// src/models/open-api-type.ts
/**
 * List of Open API types.
 */
export enum OpenApiType {
    Integer = 'integer',
    Number = 'number',
    String = 'string',
    Boolean = 'boolean',
}

/**
 * List of Open API type formats.
 */
export enum OpenApiFormat {
    Int32 = 'int32',
    Int64 = 'int64',
    Double = 'double',
    DateTime = 'date-time',
    Date = 'date'
}